feature: Dockerized multi-tenant app

+++ database/migrations/2019_09_15_000020_create_domains_table.php
@@ -18,7 +18,7 @@ class CreateDomainsTable extends Migration
         Schema::create('domains', function (Blueprint $table) {
             $table->increments('id');
             $table->string('domain', 255)->unique();
-            $table->unsignedBigInteger('tenant_id');
+            $table->unsignedInteger('tenant_id');
 
             $table->timestamps();
             $table->foreign('tenant_id')->references('id')->on('tenants')->onUpdate('cascade')->onDelete('cascade');
